Longer and Stat merged

+++ src/SalesManager.java
@@ -15,8 +15,8 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
+    public long min() {
+        long min = sales[0];
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < min) {
@@ -27,12 +27,12 @@ public class SalesManager {
         return min;
     }
 
-    public int stat() {
-        int sum = 0;
+    public long stat() {
+        long sum = 0;
         for (int i = 0; i < sales.length; i++) {
             sum += sales[i];
         }
-        int stat = 0;
+        long stat = 0;
         if (sales.length == 1) {
             stat = sum;
         }
